requirements.txt
Django==3.1
psycopg2-binary>=2.8
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
django-filter==2.4.0
django-nested-admin==3.3.2